% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.gbpca.R
\name{categorical.gbpca}
\alias{categorical.gbpca}
\title{gbpca function for categorical target variables}
\usage{
categorical.gbpca(df, versus = 8, target = "gradecat_345_36",
  best.cat = "HiB++", worst.cat = "CGO", test = "w", ssv = NULL,
  outlier_removal_ssv = TRUE)
}
\arguments{
\item{df}{Data frame to be analysed.}

\item{versus}{How many Best of the Best and Worst of the Worst do we collect? By default, we will collect 8 of each.}

\item{target}{Target varaible to be analysed. Must be categorical.
Use \code{\link{gbpca}} for continuous \code{target}.}

\item{best.cat}{The best category. The \code{versus} BOB will be selected randomly from this
category.}

\item{worst.cat}{The worst category. The \code{versus} WOW will be selected randomly from this
category.}

\item{test}{Statistical hypothesis test to be used to determine influential
process parameters. Choose between Wilcoxon Rank test (\code{"w"}, default)
and Student's t-test (\code{"t"}).}

\item{ssv}{A vector of suspected sources of variation. These are the variables
in \code{df} which we believe might have an influence on the \code{target} variable and
will be tested. If no list of \code{ssv} is provided, the test will be performed
on all numeric variables.}

\item{outlier_removal_ssv}{Logical. Should outlier removal be performed for each \code{ssv} (default: \code{TRUE})?}
}
\value{
A data frame with the following columns
\tabular{ll}{
\code{Causes} \tab Those \code{ssv} that have been found to be influential to the \code{target} variable.\cr
\code{Count} \tab The value returned by the counting method. \cr
\code{p.value} \tab The p-value of the hypothesis test performed, i.e. either of the
Wilcoxon rank test (in case \code{test = "w"}) or the t-test (if \code{test = "t"}).\cr
\code{good_lower_bound} \tab The lower bound for this \code{Cause} for good quality.\cr
\code{good_upper_bound} \tab The upper bound for this \code{Cause} for good quality.\cr
\code{bad_lower_bound} \tab The lower bound for this \code{Cause} for bad quality.\cr
\code{bad_upper_bound} \tab The upper bound for this \code{Cause} for bad quality.\cr
\code{na_removed} \tab How many missing values were in the data set for this \code{Cause}?\cr
\code{ties_best_cat} \tab How many observations fall into the best category? \cr
\code{ties_worst_cat} \tab How many observations fall into the worst category?
}
}
\description{
This function performs an good/bad - pairwise comparison analysis on a dataset and returns those parameters found to be influential.
}
\details{
We collect the Best of the Best and the Worst of the Worst
dynamically dependent on the current \code{ssv}. That means, for each \code{ssv} we first
remove all the observations with missing values for that \code{ssv} from \code{df}.
Then, based on the remaining observations, we randomly select \code{versus}
observations from the the best category (“Best of the Best”, short BOB)  and
\code{versus} observations from the worst category
(“Worst of the Worst”, short WOW). By default, we select 8 of each.
Next, we compare BOB and WOW using the the counting method and the specified
hypothesis test. If the distributions of the \code{ssv} in BOB and WOW are
significantly different, the current \code{ssv} has been identified as influential
to the \code{target} variable. An \code{ssv} is considered influential, if the test returns
a count larger/ equal to 6 and/ or a p-value of less than 0.05.
For the next \code{ssv} we again start with the entire dataset \code{df}, remove all
the observations with missing values for that new \code{ssv} and then select our
new BOB and WOW. In particular, for each \code{ssv} we might select different observations.
This dynamic selection is necessary, because in case of an incomplete data set,
if we select the same BOB and WOW for all the \code{ssv}, we might end up with many
missing values for particular \code{ssv}. In that case the hypothesis test loses
statistical power, because it is used on a smaller sample or worse, might
fail altogether if the sample size gets too small.

For those \code{ssv} determined to be significant, control bands are extracted. The rationale is:
If the value for an \code{ssv} is in the interval [\code{good_lower_bound},\code{good_upper_bound}]
the \code{target} is likely to be good. If it is in the interval
[\code{bad_lower_bound},\code{bad_upper_bound}], the \code{target} is likely to be bad.

Furthermore some summary statistics are provided: \code{na_removed} tells us
how many observations have been removed for a particular \code{ssv}. When
selecting the \code{versus} BOB/ WOW, the selection is done randomly from within
the best/ worst category, i.e. the \code{versus} BOB/ WOW are not uniquely
determined. The randomness in the selection is quantified by \code{ties_best_cat,
ties_worst_cat}, which gives the size of the best/ worst category respectively.
}
\examples{
categorical.gbpca(mtcars, target = "cyl", best.cat = "8", worst.cat = "4")

}
