% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsIris-data.R
\docType{data}
\name{resultsIris}
\alias{resultsIris}
\title{Example results data file to be used for example report generation.}
\format{A data frame as described in the documentation of \code{\link{gbpca}}.}
\usage{
resultsIris
}
\description{
This is the output of
\code{results <- gbpca(iris, target = "Sepal.Length")}
}
\keyword{datasets}
