% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validatedObsIris-data.R
\docType{data}
\name{validatedObsIris}
\alias{validatedObsIris}
\title{validatedObsIris data set}
\format{A data frame as described in the documentation of \code{\link{validate}}.}
\usage{
validatedObsIris
}
\description{
Example validation data file to be used for example report generation.
}
\details{
This is the output of

\code{
x <- validate(iris, target = "Sepal.Length",

causes = resultsIris$Causes,

results_df = resultsIris)
}

\code{validatedObsIris <- x[[1]]}
}
\keyword{datasets}
