% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationSummaryIris-data.R
\docType{data}
\name{validationSummaryIris}
\alias{validationSummaryIris}
\title{validationSummaryIris data set}
\format{A data frame as described in the documentation of \code{\link{validate}}.}
\usage{
validationSummaryIris
}
\description{
Example validation data file to be used for example report generation.
}
\details{
This is the output of

\code{
x <- validate(iris, target = "Sepal.Length",
causes = resultsIris$Causes,
results_df = resultsIris)
}

\code{validationSummaryIris <- x[[3]]}
}
\keyword{datasets}
