% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Generates report about a conducted gbpca.}
\usage{
report(df, versus = 8, target = "max17_343_36", test = "w",
  ssv = NULL, outlier_removal_target = TRUE,
  outlier_removal_ssv = TRUE, good_end = "low",
  results_path = "resultsIris.csv", validation = FALSE,
  validation_path = "Data/good_bad_validated_records.csv",
  validation_counts = "Data/good_bad_counts.csv")
}
\arguments{
\item{df}{The data frame that was analysed with \code{\link{gbpca}} or \code{\link{categorical.gbpca}}.}

\item{versus}{What value of \code{versus} was used?}

\item{target}{What \code{target} was used?}

\item{test}{Which hypothesis test was used alongside the counting method?}

\item{ssv}{Which \code{ssv} have been used in the analysis? If \code{NULL}, it
will be assumed that \code{ssv = NULL} was passed to \code{\link{gbpca}} or \code{\link{categorical.gbpca}}
and all numeric variables in \code{df} will be used.}

\item{outlier_removal_target}{Was outlier removal conducted for \code{target}?}

\item{outlier_removal_ssv}{Was outlier removal conducted for each \code{ssv}?}

\item{good_end}{Are \code{"low"} or \code{"high"} values of \code{target} good?}

\item{results_path}{Filepath to a .csv file containing results from \code{\link{gbpca}} or \code{\link{categorical.gbpca}}.}
}
\description{
Takes results from a previous gbpca and automatically generates a .pdf report
for it.
}
