% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Generates report about a conducted gbpca.}
\usage{
report(df, versus = 8, target = "Sepal.Length", type = "continuous",
  test = "w", ssv = NULL, outlier_removal_target = TRUE,
  outlier_removal_ssv = TRUE, good_outcome = "low",
  results_path = "resultsIris", validation = FALSE,
  validation_path = "validatedObsIris",
  validation_counts = "validationCountsIris",
  validation_summary = "validationSummaryIris")
}
\arguments{
\item{df}{The data frame that was analysed with \code{\link{gbpca}} or \code{\link{categorical.gbpca}}.}

\item{versus}{What value of \code{versus} was used?}

\item{target}{What \code{target} was used?}

\item{type}{Was \code{\link{gbpca}} (use \code{type = "continuous"}) or \code{\link{categorical.gbpca}} (use \code{type = "categorical"}) conducted?}

\item{test}{Which hypothesis test was used alongside the counting method?}

\item{ssv}{Which \code{ssv} have been used in the analysis? If \code{NULL}, it
will be assumed that \code{ssv = NULL} was passed to \code{\link{gbpca}} or \code{\link{categorical.gbpca}}
and all numeric variables in \code{df} will be used.}

\item{outlier_removal_target}{Was outlier removal conducted for \code{target}? If \code{type == "categorical"}
this is set to \code{FALSE} automatically.}

\item{outlier_removal_ssv}{Was outlier removal conducted for each \code{ssv}?}

\item{good_outcome}{Are \code{"low"} or \code{"high"} values of \code{target} good? Or, in
case of a categorical \code{target} the name of the best category as a string.}

\item{results_path}{Filepath to a .csv file containing results of \code{\link{gbpca}} or \code{\link{categorical.gbpca}}.}

\item{validation}{Logical. Has validation of the results been performed?}

\item{validation_path}{R object containing the validated observations, i.e. first data frame returned by \code{\link{validate}}.}

\item{validation_counts}{R object containing the counts from validation, i.e. the second data frame returned by \code{\link{validate}}.}

\item{validation_summary}{R object containing the summary of \code{validation_path}, i.e. the third data frame returned by \code{\link{validate}}.}
}
\description{
Takes results from a previous gbpca and automatically generates a html report
for it.
}
\examples{
## For continuous target
## If you want to create a new gbpca from scratch, this is the last step
## and relies on executing the other functions in this package first.
## Run gbpca
results <- gbpca(iris, target = "Sepal.Length")
## Produce regression plots
gbpca.regressions(iris, target = "Sepal.Length")
## Validate findings and store results
x <- validate(iris, target = "Sepal.Length", causes = results$Causes, results_df = results)
## Split above results
validatedObs <- x[[1]]
validationCounts <- x[[2]]
validationSummary <- x[[3]]
## Create report
report(df = iris, target = "Sepal.Length", type = "continuous", good_outcome = "low",
results_path = "results", validation = TRUE, validation_path = "validatedObs",
validation_counts = "validationCounts", validation_summary = "validationSummary")


}
